Okay for m_line, not o_line, gcc brain fart?

+++ libgeda/src/math/m_line.c
@@ -88,8 +88,8 @@ bool m_line_intersection(Line *line1, Line *line2, POINT *point)
     if (slope1 != slope2) {
 
       /* y-intercept = ordinate - slope x abscissa */
-      int b11 = line1->y[0] - (slope1 * line1->x[0]);
-      int b21 = line2->y[0] - (slope2 * line2->x[0]);
+      float b11 = line1->y[0] - (slope1 * line1->x[0]);
+      float b21 = line2->y[0] - (slope2 * line2->x[0]);
 
       /* abscissa = y-intercept2 - y-intercept1 / slope1 - slope2 */
       point->x = (b21 - b11) / (slope1 - slope2);
